SkillAPI v2.16

+++ SkillAPI/src/com/sucy/skill/mechanic/DamageMechanic.java
@@ -36,7 +36,7 @@ public class DamageMechanic implements IMechanic {
         int level = data.getSkillLevel(skill.getName());
         double damage = skill.getAttribute(DAMAGE, target, level);
         for (LivingEntity t : targets) {
-            t.setLastDamageCause(new EntityDamageByEntityEvent(player, t, EntityDamageEvent.DamageCause.CUSTOM, (double)damage));
+            t.setLastDamageCause(new EntityDamageByEntityEvent(player, t, EntityDamageEvent.DamageCause.CUSTOM, damage));
             double prevHealth = t.getHealth();
             t.damage(damage, player);
             worked = worked || prevHealth != t.getHealth();
